Update bandplot
When more than one phonon band files are ploted, the vertial line is needed as well. However, the case I'm encountering is to plot phonon bands of a material under different pressures in one figure, so I deleted the restriction of drawing vertial line for only one file. But the result figure I got showing vertial line seperatedly when I actually need the vertial lines to be merged.